Add files via upload
The purpose of this application is the part of my M.A.R.C.H.E.S concept (Military and Responder Comprehensive Health Evaluation Service), designed to
create a mental health platform for Military And Responders.

This portion uses the Beck Depression Index, a globally recognized tool to determine indication of an  depression diagnosis. This is in honor of Uncle Steve Cortez,
and my late, dear friend, Patrick Charles Berg, who suffered from schizophrenia and died from suicide in 2015. May this bring light to those seeking rescue from such a fate.